Create InstallWebsite.sh
This code is a bash script that installs and configures an Nginx web server on a Linux-based system (Debian).

The script starts with the command #!/bin/bash which is called a shebang. It tells the system to run the script using the Bash shell.

The next command sudo apt-get update updates the package list on the system to ensure that the latest version of packages are available for installation.

The command sudo apt install nginx -y installs the Nginx web server on the system. The -y flag is used to automatically confirm the installation without requiring user confirmation.

The next command sudo systemctl enable nginx starts and enables the Nginx service so that it automatically starts up whenever the system boots.

The final command sudo wget -O /var/www/html/index.nginx-debian.html https://raw.githubusercontent.com/AlessioMinerva/AWS_Terraform_Ngnix/main/webpage.html downloads a webpage and saves it as the default Nginx index page in the /var/www/html directory.